Refactoring, small adjustments
- Moved some map circle functionality to TAJMapCircle class
- Refactored TAJMapViewController class here and there
- Reduced feed radius to 250 meters
- Fixed some compiler warnings

+++ Skene/NSDate+RelativeDate.m
@@ -2,7 +2,9 @@
 //  NSDate+RelativeDate.m
 //  Skene
 //
-//  Taken from: http://stackoverflow.com/questions/4937230/relative-string-from-nsdate
+//  Adopted from: http://stackoverflow.com/questions/4937230/relative-string-from-nsdate
+//  With slight modifications:
+//  - Added (long) cast to NSDateComponents and changed string format from %d to %ld
 //
 
 #import "NSDate+RelativeDate.h"
@@ -30,31 +32,31 @@
         relativeString = @"!n the future!";
         
     } else if (delta < 1 * MINUTE) {
-        relativeString = (components.second == 1) ? @"One second ago" : [NSString stringWithFormat:@"%d seconds ago",components.second];
+        relativeString = (components.second == 1) ? @"One second ago" : [NSString stringWithFormat:@"%ld seconds ago", (long)components.second];
         
     } else if (delta < 2 * MINUTE) {
         relativeString =  @"a minute ago";
         
     } else if (delta < 45 * MINUTE) {
-        relativeString = [NSString stringWithFormat:@"%d minutes ago",components.minute];
+        relativeString = [NSString stringWithFormat:@"%ld minutes ago", (long)components.minute];
         
     } else if (delta < 90 * MINUTE) {
         relativeString = @"an hour ago";
         
     } else if (delta < 24 * HOUR) {
-        relativeString = [NSString stringWithFormat:@"%d hours ago",components.hour];
+        relativeString = [NSString stringWithFormat:@"%ld hours ago", (long)components.hour];
         
     } else if (delta < 48 * HOUR) {
         relativeString = @"yesterday";
         
     } else if (delta < 30 * DAY) {
-        relativeString = [NSString stringWithFormat:@"%d days ago",components.day];
+        relativeString = [NSString stringWithFormat:@"%ld days ago", (long)components.day];
         
     } else if (delta < 12 * MONTH) {
-        relativeString = (components.month <= 1) ? @"one month ago" : [NSString stringWithFormat:@"%d months ago",components.month];
+        relativeString = (components.month <= 1) ? @"one month ago" : [NSString stringWithFormat:@"%ld months ago", (long)components.month];
         
     } else {
-        relativeString = (components.year <= 1) ? @"one year ago" : [NSString stringWithFormat:@"%d years ago",components.year];
+        relativeString = (components.year <= 1) ? @"one year ago" : [NSString stringWithFormat:@"%ld years ago", (long)components.year];
         
     }
